Fixed typo when animating mobs - Mistakenly referenced AnimatedSprite3D instead of 2D - Slight formatting changes

+++ scripts/Main.cs
@@ -1,60 +1,60 @@
 public partial class Main : Node
 {
 
-    [Export]
-    public PackedScene MobScene { get; set; }
+	[Export]
+	public PackedScene MobScene { get; set; }
 
-    int _score;
+	int _score;
 
-    public void GameOver()
-    {
-        GetNode<Timer>("MobTimer").Stop();
-        GetNode<Timer>("ScoreTimer").Stop();
-    }
+	public void GameOver()
+	{
+		GetNode<Timer>("MobTimer").Stop();
+		GetNode<Timer>("ScoreTimer").Stop();
+	}
 
-    public void NewGame()
-    {
-        _score = 0;
+	public void NewGame()
+	{
+		_score = 0;
 
-        var player = GetNode<Player>("Player");
-        var startPos = GetNode<Marker2D>("StartPosition");
-        player.Start(startPos.Position);
+		var player = GetNode<Player>("Player");
+		var startPos = GetNode<Marker2D>("StartPosition");
+		player.Start(startPos.Position);
 
-        GetNode<Timer>("StartTimer").Start();
-    }
+		GetNode<Timer>("StartTimer").Start();
+	}
 
-    private void OnScoreTimerTimeout()
-    {
-        _score++;
-    }
+	private void OnScoreTimerTimeout()
+	{
+		_score++;
+	}
 
-    private void OnStartTimerTimeout()
-    {
-        GetNode<Timer>("MobTimer").Start();
-        GetNode<Timer>("ScoreTimer").Start();
-    }
+	private void OnStartTimerTimeout()
+	{
+		GetNode<Timer>("MobTimer").Start();
+		GetNode<Timer>("ScoreTimer").Start();
+	}
 
-    private void OnMobTimerTimeout()
-    {
-        var mob = MobScene.Instantiate<Mob>();
+	private void OnMobTimerTimeout()
+	{
+		var mob = MobScene.Instantiate<Mob>();
 
-        var mobSpawnLoc = GetNode<PathFollow2D>("Path2D/PathFollow2D");
-        mobSpawnLoc.ProgressRatio = GD.Randf();
+		var mobSpawnLoc = GetNode<PathFollow2D>("Path2D/PathFollow2D");
+		mobSpawnLoc.ProgressRatio = GD.Randf();
 
-        var direction = mobSpawnLoc.Rotation + Mathf.Pi / 2;
-        mob.Position = mobSpawnLoc.Position;
+		var direction = mobSpawnLoc.Rotation + Mathf.Pi / 2;
+		mob.Position = mobSpawnLoc.Position;
 
-        direction += (float)GD.RandRange(-Mathf.Pi / 4, Mathf.Pi / 4);
-        mob.Rotation = direction;
+		direction += (float)GD.RandRange(-Mathf.Pi / 4, Mathf.Pi / 4);
+		mob.Rotation = direction;
 
-        var velocity = new Vector2((float)GD.RandRange(150, 250), 0);
-        mob.LinearVelocity = velocity.Rotated(direction);
+		var velocity = new Vector2(GD.RandRange(150, 250), 0);
+		mob.LinearVelocity = velocity.Rotated(direction);
 
-        AddChild(mob);
-    }
+		AddChild(mob);
+	}
 
-    public override void _Ready()
-    {
-        NewGame();
-    }
+	public override void _Ready()
+	{
+		NewGame();
+	}
 }
